Document usage of can-event/lifecycle/lifecycle
